better abstraction for model and query builder classes
